# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Soccer Prediction API
  version: 0.4.0
  description: API for the soccer prediction web application.

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /api/users/me:
    get:
      summary: Get current authenticated user
      tags: [Users]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Not authenticated

  /api/auth/logout:
    post:
      summary: Log out the current user
      tags: [Auth]
      responses:
        '302':
          description: Redirect to frontend after logout

  /api/leagues:
    get:
      summary: List current user's leagues
      tags: [Leagues]
      responses:
        '200':
          description: List of leagues the user has joined
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeagueSummaryDto'
        '401':
          description: Not authenticated

  /api/leagues/join:
    post:
      summary: Join a league by code
      tags: [Leagues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinLeagueRequest'
      responses:
        '201':
          description: Successfully joined league
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueSummaryDto'
        '200':
          description: Already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueSummaryDto'
        '404':
          description: Invalid join code
        '401':
          description: Not authenticated

  /api/leagues/{id}:
    get:
      summary: Get league detail
      tags: [Leagues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: League detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDto'
        '403':
          description: Not a league member
        '401':
          description: Not authenticated

  /api/leagues/{id}/fixtures:
    get:
      summary: Get fixtures for a league
      tags: [Leagues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of fixtures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixtureDto'
        '403':
          description: Not a league member
        '401':
          description: Not authenticated

  /api/leagues/{id}/standings:
    get:
      summary: Get standings for a league
      tags: [Leagues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: League standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StandingDto'
        '403':
          description: Not a league member
        '401':
          description: Not authenticated

  /api/predictions/{fixtureId}:
    put:
      summary: Create or update a match prediction
      tags: [Predictions]
      parameters:
        - name: fixtureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        '200':
          description: Prediction saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionDto'
        '400':
          description: Kickoff has passed
        '403':
          description: Not a league member
        '404':
          description: Fixture not found
        '401':
          description: Not authenticated

  /api/leagues/{id}/predictions/me:
    get:
      summary: Get current user's predictions for a league
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of predictions with fixture details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PredictionDto'
        '401':
          description: Not authenticated

  /api/leagues/{id}/players:
    get:
      summary: Get players for a league (for top scorer pick)
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerDto'
        '401':
          description: Not authenticated

  /api/leagues/{id}/top-scorer-pick:
    get:
      summary: Get current user's top scorer pick
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Top scorer pick
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopScorerPickDto'
        '204':
          description: No pick set yet
        '401':
          description: Not authenticated
    put:
      summary: Set top scorer pick
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopScorerPickRequest'
      responses:
        '200':
          description: Pick saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopScorerPickDto'
        '400':
          description: Picks are locked
        '403':
          description: Not a league member
        '401':
          description: Not authenticated

  /api/leagues/{id}/league-winner-pick:
    get:
      summary: Get current user's league winner pick
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: League winner pick
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueWinnerPickDto'
        '204':
          description: No pick set yet
        '401':
          description: Not authenticated
    put:
      summary: Set league winner pick
      tags: [Predictions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeagueWinnerPickRequest'
      responses:
        '200':
          description: Pick saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueWinnerPickDto'
        '400':
          description: Picks are locked
        '403':
          description: Not a league member
        '401':
          description: Not authenticated

  /api/admin/leagues:
    get:
      summary: List all leagues (admin)
      tags: [Admin]
      responses:
        '200':
          description: All leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeagueDto'
        '403':
          description: Not an admin
    post:
      summary: Create a league (admin)
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeagueRequest'
      responses:
        '201':
          description: League created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDto'
        '403':
          description: Not an admin

  /api/admin/leagues/{id}:
    put:
      summary: Update league settings (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeagueRequest'
      responses:
        '200':
          description: League updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDto'
        '403':
          description: Not an admin

  /api/admin/leagues/search:
    get:
      summary: Search API-Football leagues (admin)
      tags: [Admin]
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API-Football league search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiFootballLeague'
        '403':
          description: Not an admin

  /api/admin/leagues/{id}/teams:
    get:
      summary: List teams in a league (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Teams in the league
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamDto'
        '403':
          description: Not an admin
    post:
      summary: Add team to league (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDto'
        '400':
          description: League is not manual mode
        '403':
          description: Not an admin

  /api/admin/teams/{id}:
    put:
      summary: Update team (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDto'
        '403':
          description: Not an admin
    delete:
      summary: Remove team (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Team deleted
        '403':
          description: Not an admin

  /api/admin/leagues/{id}/fixtures:
    get:
      summary: List fixtures in a league (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fixtures in the league
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixtureDto'
        '403':
          description: Not an admin
    post:
      summary: Create fixture (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFixtureRequest'
      responses:
        '201':
          description: Fixture created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureDto'
        '400':
          description: League is not manual mode
        '403':
          description: Not an admin

  /api/admin/fixtures/{id}:
    put:
      summary: Update fixture (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFixtureRequest'
      responses:
        '200':
          description: Fixture updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureDto'
        '403':
          description: Not an admin
    delete:
      summary: Remove fixture (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Fixture deleted
        '403':
          description: Not an admin

  /api/admin/fixtures/{id}/result:
    put:
      summary: Enter match result (admin)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterResultRequest'
      responses:
        '200':
          description: Result entered, scoring triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureDto'
        '403':
          description: Not an admin

components:
  schemas:
    UserDto:
      type: object
      required: [id, email, displayName, role, isActive]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        pictureUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, ADMIN]
        isActive:
          type: boolean

    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object

    LeagueDto:
      type: object
      required: [id, name, mode, season, joinCode, exactScorePoints, correctOutcomePoints, wrongPredictionPoints, topScorerBonus, leagueWinnerBonus, memberCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        mode:
          type: string
          enum: [MANUAL, API_SYNCED]
        apiLeagueId:
          type: integer
          nullable: true
        season:
          type: integer
        joinCode:
          type: string
        exactScorePoints:
          type: integer
        correctOutcomePoints:
          type: integer
        wrongPredictionPoints:
          type: integer
        topScorerBonus:
          type: integer
        leagueWinnerBonus:
          type: integer
        memberCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time

    LeagueSummaryDto:
      type: object
      required: [id, name, season, memberCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        season:
          type: integer
        memberCount:
          type: integer
          format: int64

    CreateLeagueRequest:
      type: object
      required: [name, season]
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [MANUAL, API_SYNCED]
          default: API_SYNCED
        apiLeagueId:
          type: integer
          nullable: true
        season:
          type: integer
        exactScorePoints:
          type: integer
          default: 3
        correctOutcomePoints:
          type: integer
          default: 1
        wrongPredictionPoints:
          type: integer
          default: 0
        topScorerBonus:
          type: integer
          default: 10
        leagueWinnerBonus:
          type: integer
          default: 10

    UpdateLeagueRequest:
      type: object
      properties:
        name:
          type: string
        exactScorePoints:
          type: integer
        correctOutcomePoints:
          type: integer
        wrongPredictionPoints:
          type: integer
        topScorerBonus:
          type: integer
        leagueWinnerBonus:
          type: integer

    JoinLeagueRequest:
      type: object
      required: [joinCode]
      properties:
        joinCode:
          type: string

    TeamDto:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        countryCode:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        groupName:
          type: string
          nullable: true

    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        countryCode:
          type: string
        groupName:
          type: string

    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        countryCode:
          type: string
        groupName:
          type: string

    FixtureDto:
      type: object
      required: [id, homeTeam, awayTeam, kickoff, status, matchday]
      properties:
        id:
          type: string
          format: uuid
        apiFixtureId:
          type: integer
          nullable: true
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeTeamLogo:
          type: string
          nullable: true
        awayTeamLogo:
          type: string
          nullable: true
        kickoff:
          type: string
          format: date-time
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        status:
          type: string
          enum: [SCHEDULED, LIVE, FINISHED, POSTPONED, CANCELLED]
        round:
          type: string
          nullable: true
        matchday:
          type: integer

    CreateFixtureRequest:
      type: object
      required: [homeTeamId, awayTeamId, kickoff, matchday]
      properties:
        homeTeamId:
          type: string
          format: uuid
        awayTeamId:
          type: string
          format: uuid
        kickoff:
          type: string
          format: date-time
        round:
          type: string
        matchday:
          type: integer

    UpdateFixtureRequest:
      type: object
      properties:
        homeTeamId:
          type: string
          format: uuid
        awayTeamId:
          type: string
          format: uuid
        kickoff:
          type: string
          format: date-time
        round:
          type: string
        matchday:
          type: integer

    EnterResultRequest:
      type: object
      required: [homeScore, awayScore]
      properties:
        homeScore:
          type: integer
        awayScore:
          type: integer

    StandingDto:
      type: object
      required: [id, teamName, rank, points, played, won, drawn, lost, goalsFor, goalsAgainst, goalDiff]
      properties:
        id:
          type: string
          format: uuid
        apiTeamId:
          type: integer
          nullable: true
        teamName:
          type: string
        teamLogo:
          type: string
          nullable: true
        groupName:
          type: string
          nullable: true
        rank:
          type: integer
        points:
          type: integer
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        goalDiff:
          type: integer

    PredictionDto:
      type: object
      required: [id, fixtureId, homeScore, awayScore, fixtureHomeTeam, fixtureAwayTeam, fixtureKickoff, fixtureStatus, matchday]
      properties:
        id:
          type: string
          format: uuid
        fixtureId:
          type: string
          format: uuid
        homeScore:
          type: integer
        awayScore:
          type: integer
        pointsEarned:
          type: integer
          nullable: true
        fixtureHomeTeam:
          type: string
        fixtureAwayTeam:
          type: string
        fixtureHomeScore:
          type: integer
          nullable: true
        fixtureAwayScore:
          type: integer
          nullable: true
        fixtureKickoff:
          type: string
          format: date-time
        fixtureStatus:
          type: string
        matchday:
          type: integer

    PredictionRequest:
      type: object
      required: [homeScore, awayScore]
      properties:
        homeScore:
          type: integer
        awayScore:
          type: integer

    TopScorerPickDto:
      type: object
      required: [id, playerName]
      properties:
        id:
          type: string
          format: uuid
        playerName:
          type: string
        apiPlayerId:
          type: integer
          nullable: true
        pointsEarned:
          type: integer
          nullable: true

    TopScorerPickRequest:
      type: object
      required: [playerName]
      properties:
        playerName:
          type: string
        apiPlayerId:
          type: integer

    LeagueWinnerPickDto:
      type: object
      required: [id, teamName]
      properties:
        id:
          type: string
          format: uuid
        teamName:
          type: string
        apiTeamId:
          type: integer
          nullable: true
        pointsEarned:
          type: integer
          nullable: true

    LeagueWinnerPickRequest:
      type: object
      required: [teamName]
      properties:
        teamName:
          type: string
        apiTeamId:
          type: integer

    PlayerDto:
      type: object
      required: [apiPlayerId, name, apiTeamId]
      properties:
        apiPlayerId:
          type: integer
        name:
          type: string
        photoUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        apiTeamId:
          type: integer

    ApiFootballLeague:
      type: object
      properties:
        leagueId:
          type: integer
        name:
          type: string
        country:
          type: string
        logo:
          type: string
          nullable: true
        seasons:
          type: array
          items:
            type: integer